+++ test/unit/framework/fixtures/log_dir/collector/biosdatamodel.json
@@ -1,0 +1,3 @@
+{
+  "bios_version": "M17"
+}
